Add a basic UISearchController to the FollowersListVC
There is bug with the search controller that is not shown when the username has a few followers (2, 4, 5, etc. When FollowerListVC doesn't have to scroll to show all the followers). Also there's other bug that the search bar is shown when you search a user and then go back and immediately search for the same user. Until you scroll and scroll back is when the search bar appears.