Calling the getFavorites() method now that it can get the favorites array and fills the tableview or shows the alert.
This method is called in viewWillAppear(_:) because viewDidLoad() is called once, the favoritesListVC is not destroyed when changing between tabs, so in the edge case, what if we launch the app and we go to favorites screen and we don't have favorites yet. And then we get back to search and add a username to favorites. If we move to favorites screen the viewDidLoad is not called again, the method that is calles is viewWillAppear()