There is a bug now with the downloadImage(from:completed) refactor made a couple of commits ago. When the network is too slow it will get a nil UIImage in the completed handler and the collection view will load an empty avatar images instead of the placeholder image. My own fix, instead of sending nil, send the placeholder image
Other solution and who found the bug
Comment from: Theodore Hecht
In my refactored network code for downloadImage, a bug existed where network errors would result in the avatarImageView's default placeholder image being overwritten to a blank/nil image. This happens because on a network error, the network manager passes nil to the downloadImage's completion handler, which then gets set to the avatarImageView's image property at the calling site, regardless of whether or not the UIImage? contained a UIImage or nil. This results in an overwritten avatarImageView even in the case of a nil/blank image.

Here are two possible fixes:

(1) At every calling site for downloading an image, in the completion handler which receives a UIImage?, make sure to unwrap the optional to ensure it isn't nil before setting avatarImageView.image = image.

(2) We could instead refactor the NetworkManager's download image completion parameter to pass a non-optional UIImage, calling the completion handler with a placeholder image UIImage instead of nil.